Create class LSystemRenderer
Will render a mesh based on a LSystem
